legacy_versions/v1-01/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
PyJWT==2.9.0
requests==2.32.3
gunicorn==21.2.0

# Realtime via Socket.IO
Flask-SocketIO==5.3.6
gevent==24.2.1
gevent-websocket==0.10.1
psutil==6.0.0
redis==5.0.1

